libhbp/requirements.txt
black
build
colorama
mysql-connector-python
no_implicit_optional
mypy
pipreqs
